move discouraged functions to discrete package
