+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/symmetric/decrypt.kt
@@ -10,30 +10,6 @@ import at.asitplus.signum.supreme.mac.mac
 import org.kotlincrypto.SecureRandom
 import kotlin.jvm.JvmName
 
-/**
- * Encrypts [data] and automagically generates a fresh IV if required by the cipher.
- * This is the method you want to use, as it generates a fresh IV, if the underlying cipher requires an IV.
- * * [authenticatedData] = _Additional Authenticated Data_
- *
- * It is safe to discard the reference to [authenticatedData], as both IV and AAD will be added to any [at.asitplus.signum.indispensable.symmetric.Ciphertext.Authenticated] resulting from an encryption.
- *
- * @return [KmmResult.success] containing a [at.asitplus.signum.indispensable.symmetric.Ciphertext.Authenticated] if valid parameters were provided or [KmmResult.failure] in case of
- * invalid parameters (e.g., key or IV length)
- */
-@JvmName("encryptAuthenticated")
-fun <A : CipherKind.Authenticated, I : Nonce> SymmetricKey<A, I>.encrypt(
-    data: ByteArray,
-    authenticatedData: ByteArray? = null
-): KmmResult<SealedBox<A, I, SymmetricEncryptionAlgorithm<A, I>>> = catching {
-    Encryptor(
-        algorithm,
-        secretKey,
-        if (this is WithDedicatedMac) dedicatedMacKey else secretKey,
-        null,
-        authenticatedData,
-    ).encrypt(data) as SealedBox<A, I, SymmetricEncryptionAlgorithm<A, I>>
-}
-
 
 /**
  * Attempts to decrypt this ciphertext (which also holds IV, and in case of an authenticated ciphertext, AAD and auth tag) using the provided [key].
