+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/symmetric/encrypt.kt
@@ -15,21 +15,6 @@ import at.asitplus.signum.indispensable.symmetric.SymmetricKey.WithDedicatedMac
 import kotlin.jvm.JvmName
 
 
-@HazardousMaterials
-@JvmName("encryptWithIV")
-fun <A : CipherKind> SymmetricKey<A, Nonce.Required>.encrypt(
-    iv: ByteArray,
-    data: ByteArray
-): KmmResult<SealedBox.WithNonce<A, SymmetricEncryptionAlgorithm<A, Nonce.Required>>> = catching {
-    Encryptor(
-        algorithm,
-        secretKey,
-        if (this is WithDedicatedMac) dedicatedMacKey else secretKey,
-        iv,
-        null,
-    ).encrypt(data) as SealedBox.WithNonce<A, SymmetricEncryptionAlgorithm<A, Nonce.Required>>
-}
-
 @JvmName("encryptWithAutoGenIV")
 fun <A : CipherKind, I : Nonce> SymmetricKey<A, I>.encrypt(
     data: ByteArray
@@ -43,27 +28,27 @@ fun <A : CipherKind, I : Nonce> SymmetricKey<A, I>.encrypt(
     ).encrypt(data) as SealedBox<A, I, SymmetricEncryptionAlgorithm<A, I>>
 }
 
+
 /**
- * Encrypts [data] using a specified IV. Check yourself, before you really, really wreck yourself!
- * * [iv] =  _Initialization Vector_; **NEVER EVER RE-USE THIS!**
+ * Encrypts [data] and automagically generates a fresh IV if required by the cipher.
+ * This is the method you want to use, as it generates a fresh IV, if the underlying cipher requires an IV.
  * * [authenticatedData] = _Additional Authenticated Data_
  *
- * It is safe to discard the reference to [iv] and [authenticatedData], as both will be added to any [at.asitplus.signum.indispensable.symmetric.Ciphertext.Authenticated] resulting from an encryption.
+ * It is safe to discard the reference to [authenticatedData], as both IV and AAD will be added to any [at.asitplus.signum.indispensable.symmetric.Ciphertext.Authenticated] resulting from an encryption.
  *
  * @return [KmmResult.success] containing a [at.asitplus.signum.indispensable.symmetric.Ciphertext.Authenticated] if valid parameters were provided or [KmmResult.failure] in case of
  * invalid parameters (e.g., key or IV length)
  */
-@HazardousMaterials
-fun <A : CipherKind.Authenticated> SymmetricKey<A, Nonce.Required>.encrypt(
-    iv: ByteArray,
+@JvmName("encryptAuthenticated")
+fun <A : CipherKind.Authenticated, I : Nonce> SymmetricKey<A, I>.encrypt(
     data: ByteArray,
     authenticatedData: ByteArray? = null
-): KmmResult<SealedBox.WithNonce<A, SymmetricEncryptionAlgorithm<A, Nonce.Required>>> = catching {
+): KmmResult<SealedBox<A, I, SymmetricEncryptionAlgorithm<A, I>>> = catching {
     Encryptor(
         algorithm,
         secretKey,
         if (this is WithDedicatedMac) dedicatedMacKey else secretKey,
-        iv,
+        null,
         authenticatedData,
-    ).encrypt(data) as SealedBox.WithNonce<A, SymmetricEncryptionAlgorithm<A, Nonce.Required>>
+    ).encrypt(data) as SealedBox<A, I, SymmetricEncryptionAlgorithm<A, I>>
 }
